change uint64 to uint

+++ model/content.go
@@ -6,7 +6,7 @@ import (
 
 type Content struct {
 	gorm.Model
-	CollectionID  uint64         `gorm:"not null"`
+	CollectionID  uint           `gorm:"not null"`
 	Collection    Collection     `gorm:"foreignKey:CollectionID"`
 	ContentValues []ContentValue `gorm:"foreignKey:ContentID;references:ID"`
 }
